add support file for athena, crazyflie, linux, and macos
